https://www.reddit.com/r/reinforcementlearning/comments/8gdtei/decoupling_dynamics_and_reward_for_transfer/dz4lgt7/?st=jhajazzc&sh=4d46ef51
"> I watched the video, it was great, but I didn't pick up that part. Care to expand on that?

Think what happens if you have a sequence a million steps long. What MC would do is try to predict the reward at the end regardless of the intermediate steps. As a learning task it would be something analogous to you trying to predict your exact location tens years from now. It cannot be done, and the whole setup is absurd.

TD learning on the other hand would split the huge task of predicting many steps ahead into predicting short subsequences and chain them together via bootstrapping.

MC works for me right now since the poker game I am trying it now does not have that many steps in each hand, but if I tried predicting the end result of the whole game or even multiple games it would not work.

> Huh that's interesting.

It started with the [Natural Neural Networks](https://arxiv.org/abs/1507.00210) paper. Note that the algorithm presented there only does whitening every once in a while otherwise it would be too slow. This does make the learning less stable since it degenerates into SGD.

More recently there has been an extension on it - [A Neural Network model with Bidirectional Whitening](https://arxiv.org/abs/1704.07147) while also whitens the gradients during the backward pass.

Something interesting might come out of this line of research if it could be made less computationally intensive, but even as I made the last reply to you I started having doubts. If an algorithm requires natural gradient updates just to work then there is probably something wrong with it.

I had a really good idea last night how to do TD-style learning without the huge disadvantage of having to do explicit bootstrapping directly via the cost function.

I reasoned out how to associate actions to states.

Even a month ago I had the intuition that both MC and TD updates would be wrong (among some other intuitions that have been now washed away by experience.) The basic idea was to avoid the using the cost function to propagate information and instead do it completely through recurrent dynamics.

Consider how in Q learning uses the cost function as a channel for propagating values across time in feedforward nets. That definitely feels amiss from a design perspective. Not only is it deeply unstable and requires hacks such as target networks to work, but also it does not at all take advantage of the fact that recurrent networks exist and have perfectly workable temporal channels through which credit assignment might be done.

That last insight is the starting point for the idea. The issue with it is just how to propagate the rewards without doing MC or explicit TD/Q learning updates.

It would not be enough to just add recurrent connections and predict the reward directly. In poker for example, the problem that would optimize would be to maximize reward only in the final step.

In order to do full propagation then what would be needed is to propagate gradients from the input (the state) into the previous step's output (the action.)

The issue with that is that it is impossible. Since the input is given by the game itself, it is not directly related to the action and there is nothing to propagate through.

The are two false ways you could think to force the association which are more intuitive than the correct way.

1) Give the previous action as a part of the input. That is rather than having the policy network be `s -> a`, make it `s * a -> a`. Since actions are differentiable that would make a channel into the previous step for the gradients, but there is an issue with this.

In poker for example, the action you took is entirely predictable from the state which falsifies the whole idea. If the network decided to ignore the action part of the input there would be no reason to ever pay attention to it again and it would block the gradients. At most one could hope that the gradient would flow through.

So this idea is a bust.

2) Somehow use the value or the action (which are just scalars) and project them into a vector that could then be used to multiply the state. The goal would be to make it like an attention operation. It would then be possible to propagate the gradients through that.

This particular idea is absurd for the simple reason that the only operation which even remotely makes sense to use as a projection would be to a vector of all ones. But with multiplication the zero entries in the input would propagate no gradient - does that make sense at all? Probably not.

But this thought process actually leads closer to the truth.

3) By process of elimination the only correct option here could be to project a scalar action which is just a probability scalar to a vector of all zeroes and then add them to the next step's input.

This is highly unintuitive because in normal math if you have something like `x=0; y=x+z` then you would expect that `y=z` to hold. In differentiable programming the above would be the association operation and would create a link from `y` to `x`.

Projecting to and then adding zeroes, ugh. I really couldn't have thought of this earlier. It really is the last thing that comes to mind.

The second key part of the plan is how to avoid the need to do explicit bootstrapping with not just a single episodic return, but across a whole lifetime.

The way to do it is just to use policy gradients.

An interesting aspect of the PG algorithm is that it is the equivalent of global mutation in regular programming. It makes it on-policy, but it should also make it unnecessary to do that torturous explicit bootstrapping process. It really makes it quite easy to propagate rewards - imagine the policy as a counter and on the forward pass just push it up or down. The backward pass (which is linear) will take care of the rest.

One note to add on policy gradients is to not do the inane thing of using cross entropy as the cost function. Instead the output's adjoint should be set to the reward directly. The issue with cross entropy is that if the action probability is 0.1 and the reward is 0.1 then the resulting gradient will be zero which is something that should not be allowed to happen as it would bias the training. Cross entropy is for prediction, not mutation and the two should not be confused.

---

This outlines my idea. I haven't tested it yet, but hopefully the recipe is correct. I will try it out in the following days. I really do have a lot of hope for this. If it works then that means I now understand a fundamental part of deep RL that is not visible to the rest of the practitioners.

The two ingredients which are action-state association and policy gradients exactly take advantage of all the strengths of recurrent nets. The reason PG have such high variance/overfit is because they do MC updates, so it stands to reason that here due to the implicit bootstrapping they should be closer to the ideal low bias and low variance algorithm."

Just look at this discussion. I should be reading papers, but I am completely consumed by this idea. My brain feels like it is on fire.

I'll start work on it when I clear out the backlog and get a proper night's sleep.

It has to be this, it just has to.