Right now I am thinking what I should do next.
I want to resume programming that is for sure, but first I need to clean up the mess that I made.

I really did not intend for so many details of the player to spill into the actual game. That is a huge oversight. And the game is actually a lot more complicated than I intended it to be because of that.

Because of that poor design I had to remove the MC section when I moved to Q learning. That is just nasty. Aren't I supposed to compare different methods? The quality of my research if it can be called that is quite low.

Also now that I need to pass the previous action into the player how should I do that? I'd have to modify the game once more.

And all those heapifications and indivs?

No, that is not what I had in mind. The current state is just unacceptable.

Before I start work on PG or anything else, I really need to reconsider the design of the whole thing from the ground up. Then I'll do the PG idea.

For the DQN, I do not have an interest in finishing it. I'll leave it in a separate branch for that reason just in case I need the code later.

Let me clear out all of the crud.