I am really feeling inspired, but let me stop so I can eat.
I am imagining and in showdown I really made a huge error by not using arrays to get active players and winners. Not doing that makes the whole function really a lot more complicated than it should be. I'll get rid of functional purity in the game.

Last time my real problem was that I merged the game states with the player internal state have been paying the price for it until now. The move to functional purity was a move so I could get better separation of concerns, but even as I wrote the thing I got the sense that I was not getting quite the feedback that I wanted. That feeling turned out to be true.

Though I am supposedly an expert programmer now, the quality of the RL parts and the game is surprisingly low.