```     inl reply_random {name} =         inl rnd = Random()         inl _ {fold call raise} ->             match rnd.next(0i32,5i32) with             | 0i32 -> fold ()             | 1i32 -> call ()             | _ -> raise 0 ```
Here is the reply for random. Yeah, this is what it should be.

```
    inl reply_random {name} =
        inl rnd = Random()
        inl reply = inl _ _ {fold call raise} ->
            match rnd.next(0i32,5i32) with
            | 0i32 -> fold {state=(); bet=()}
            | 1i32 -> call {state=(); bet=()}
            | _ -> raise {state=(); bet=()} 0

        inl rec trace = function
            | .add_bet _ -> trace
            | .add_reward _ -> trace
            | .process -> ()

        {reply name trace state=()}
```

This is what it was last time. It is no wonder I hated this shit.

It made over player related to every other player. No way was that acceptable.

Let me do the MC player and I will call it a day.

The players are actually rather long for some reason I do not understand.