Fixed a bunch of errors and now I have the tabular player back in action again.
Phew.

I do not think I have enough steam to really start work on the PG methods today. But before I do that, let me completely cut out the RL segments.

I'll forget that it ever existed and rebuilt it with a clear mind.

I'll start off with vanilla PG and then move to testing out my particular idea.

If it fails, I won't give up. Instead I will just implement differentiable plasticity and try it again.

I am 100% sure that the existing costs are wrong, but I am much less certain that my idea is right.

If it turns out to be wrong then probably...differentiable plasticity will be the answer.

That is simply how things must go.

For training a controller, the PG mutation is absolutely a key aspect of the puzzle that needs to be taken advantage of. Similarly for recurrent connections.

Even without recurrent nets, I will get an answer that I seek. I should just train a linear controller and then try associative connections that I am thinking of.

If that fails and it could, then I suppose I could also try recurrent nets since it won't be hard, but it won't work on that either.

After that I'll be able to sweep this possibility off the table. In that case, some kind of novel architecture like the DP one will be the true key.

Well, let me do some cleaning and I will call it a day.