Ready for action.
I am thinking if both the action state association and the RNNs fail, what I am going to do is place the blame on backprop itself.

There was a talk I posted on the ML sub a while back where they used a biologically inspired algorithm to do the updates. I am going to look into it in that case.

This is such a challenging part of the whole path. I absolutely must find a way to stabilize the updates.

I do not think the general RL framework is wrong. Rather the deep RL algorithm are lacking in some fundamental manner and I need to figure out where exactly.

If I can move on from this step, then I will be get to the phase of using RL for interesting things.

I might be stuck on this for a while.

The one thing I absolutely have control over is the design of the language and the library and the game. While I collect the pieces that will eventually clear this phase, I will make sure that I attain the absolute pinnacle of design.

It won't be long now. I already tried MC and DQ learning. Only PG is left. After I try it out, I will bring out those other pieces and put them into the library as I might need them later.