Let me make sure it is absolutely the same.
```
Starting timing for: iteration 0
Winrate is 700 and 300 out of 1000.
The time was 00:00:00.7515335 for: iteration 0
Starting timing for: iteration 1
Winrate is 745 and 255 out of 1000.
The time was 00:00:00.8529026 for: iteration 1
Starting timing for: iteration 2
Winrate is 801 and 199 out of 1000.
The time was 00:00:00.9328471 for: iteration 2
Starting timing for: iteration 3
Winrate is 781 and 219 out of 1000.
The time was 00:00:00.9746971 for: iteration 3
Starting timing for: iteration 4
Winrate is 775 and 225 out of 1000.
The time was 00:00:00.9964647 for: iteration 4
Starting timing for: iteration 5
Winrate is 766 and 234 out of 1000.
The time was 00:00:01.0280378 for: iteration 5
Starting timing for: iteration 6
Winrate is 806 and 194 out of 1000.
The time was 00:00:00.9962889 for: iteration 6
Starting timing for: iteration 7
Winrate is 779 and 221 out of 1000.
The time was 00:00:01.0090120 for: iteration 7
Starting timing for: iteration 8
Winrate is 806 and 194 out of 1000.
The time was 00:00:01.0539042 for: iteration 8
Starting timing for: iteration 9
Winrate is 797 and 203 out of 1000.
The time was 00:00:01.0736701 for: iteration 9
```

This is what I got with the new one. Wow.

```
    inl range = 32 // TODO: In debugging mode.
    inl scale = max_stack_size / num_players |> to float64
```

The only thing I did was set the range and scale to what it was in the previous branch.

There is not much I can do about scale, but range is just suspicious to me.

I am going to print it out to make sure that it is not getting cut off or something.