Let me stop here. I am done with the two players. After I get back I will fix `CudaRandom` and try making the layer recurrent. After that I'll replace the MC updates in `player_pg` with direct ones.
There should be a way to make this somehow work better than the baseline with learning rate set to 0 which give me a 33% win rate. Actually no, the real baseline should be compared to only with the last layer trained.

It will be tricky to get this to work with a recurrent layer as the decoder. That gave me trouble in the char-RNN. Here it absolutely must be that though...maybe. I'll test it out.

There is not need to make wild assumptions here. I am finally getting a handle on this anyway design wise.