"https://arxiv.org/pdf/1803.04831.pdf Independently Recurrent Neural Network (IndRNN) - Building A Longer and Deeper RNN
Let me read this paper.

1:05pm. https://github.com/batzner/indrnn/blob/master/ind_rnn_cell.py

This paper is amazing. Here is the implementation. It seems I only need to clip the absolute value by `2 ** (1 / timesteps)`.

That means between `[1,2]` depending on the number of timesteps.

1:35pm. Ok, that paper was really interesting. I am not sure how to determine the clipping ranges so I asked on the ML sub. It really distracted me.

The results for this are particularly good. It is quite impressive. Should I implement differentiable plasticity or this...well, it does not matter. Let me just get on with this."

Let me resume. Before I got distracted by this, what exactly was my plan?

...To fix CudaRandom so it takes odd number of elements. After that I'll try replacing the linear layer in the PG implementation with a recurrent one.

In addition to that, I'll make a different implementation of PG where the trace is not MC, but direct instead.

I need to test what PG does in that regime.