"Let me try the mutator.
2:40pm.

```
Starting timing for: iteration 0
Winrate is 651 and 349 out of 1000.
The time was 00:00:03.3044891 for: iteration 0
Starting timing for: iteration 1
Winrate is 596 and 404 out of 1000.
The time was 00:00:03.1420646 for: iteration 1
Starting timing for: iteration 2
Winrate is 614 and 386 out of 1000.
The time was 00:00:03.1572500 for: iteration 2
Starting timing for: iteration 3
Winrate is 626 and 374 out of 1000.
The time was 00:00:03.2755775 for: iteration 3
Starting timing for: iteration 4
Winrate is 632 and 368 out of 1000.
The time was 00:00:03.0672753 for: iteration 4
Starting timing for: iteration 5
Winrate is 634 and 366 out of 1000.
The time was 00:00:03.2034132 for: iteration 5
Starting timing for: iteration 6
Winrate is 648 and 352 out of 1000.
The time was 00:00:03.1171636 for: iteration 6
Starting timing for: iteration 7
Winrate is 651 and 349 out of 1000.
The time was 00:00:03.1385442 for: iteration 7
Starting timing for: iteration 8
Winrate is 622 and 378 out of 1000.
The time was 00:00:03.2321007 for: iteration 8
Starting timing for: iteration 9
Winrate is 629 and 371 out of 1000.
The time was 00:00:03.2159659 for: iteration 9
```

Oh this is really good. This is the miln with the mutator update. Works much better than PG, and a bit better than the standard thing.

What I am especially thrilled is how much it gets in the first iteration. I think after the first iteration it essentially just gets stuck. If it gets locked in to a good strategy it will get stuck there.

I want to make this work well.

In order to encourage exploration...how about I shift the rewards? For MC updates that is a bad idea, but for mutators that should be standard."