No, it is nothing. Let me bring back the rules player.
Edit:

```
Starting timing for: iteration 0
Winrate is 756 and 244 out of 1000.
The time was 00:00:00.0238977 for: iteration 0
Starting timing for: iteration 1
Winrate is 754 and 246 out of 1000.
The time was 00:00:00.0031483 for: iteration 1
Starting timing for: iteration 2
Winrate is 743 and 257 out of 1000.
The time was 00:00:00.0030046 for: iteration 2
Starting timing for: iteration 3
Winrate is 794 and 206 out of 1000.
The time was 00:00:00.0029656 for: iteration 3
Starting timing for: iteration 4
Winrate is 773 and 227 out of 1000.
The time was 00:00:00.0029926 for: iteration 4
Starting timing for: iteration 5
Winrate is 796 and 204 out of 1000.
The time was 00:00:00.0030801 for: iteration 5
Starting timing for: iteration 6
Winrate is 780 and 220 out of 1000.
The time was 00:00:00.0031538 for: iteration 6
Starting timing for: iteration 7
Winrate is 746 and 254 out of 1000.
The time was 00:00:00.0031586 for: iteration 7
Starting timing for: iteration 8
Winrate is 775 and 225 out of 1000.
The time was 00:00:00.0030175 for: iteration 8
Starting timing for: iteration 9
Winrate is 763 and 237 out of 1000.
The time was 00:00:00.0031343 for: iteration 9
```

This player is not bad against the random player. Let me test it against the tabular one.