"5:40pm. It would be worth it to try making a sampling square player. Let me give it a shot.
```
        sampling_square = inl temp x s ->
            inl dim_a, dim_b = primal x .dim
            inl batch_size = HostTensor.span dim_a
            assert (batch_size = 1) "Only a single dimension for now."

            inl p = softmax temp (primal x) s
            inl a = sample_body p s

            a,inl (reward: float64) ->
                inl batch_size = to float batch_size
                inl reward = to float reward
                inl a,x_p,x_a = Tuple.map to_dev_tensor (a,primal x,adjoint x)
                s.CudaKernel.iter () (inl j ->
                    inl a = a j .get
                    inl x_p,x_a = Tuple.map (inl x -> x j a) (x_p,x_a)
                    inl v = x_p.get
                    x_a.set (x_a.get + square_bck (v, reward) / batch_size)
                    ) (fst x.dim)
```

For sampling I had to copy paste it.

Well, let me give it a shot.

```
Starting timing for: iteration 0
Winrate is 196 and 804 out of 1000.
The time was 00:00:01.6044783 for: iteration 0
Starting timing for: iteration 1
Winrate is 216 and 784 out of 1000.
The time was 00:00:01.5998325 for: iteration 1
Starting timing for: iteration 2
Winrate is 222 and 778 out of 1000.
The time was 00:00:01.6993891 for: iteration 2
Starting timing for: iteration 3
Winrate is 225 and 775 out of 1000.
The time was 00:00:01.6204987 for: iteration 3
Starting timing for: iteration 4
Winrate is 243 and 757 out of 1000.
The time was 00:00:01.6621881 for: iteration 4
Starting timing for: iteration 5
Winrate is 234 and 766 out of 1000.
The time was 00:00:01.5895660 for: iteration 5
Starting timing for: iteration 6
Winrate is 226 and 774 out of 1000.
The time was 00:00:01.6248715 for: iteration 6
Starting timing for: iteration 7
Winrate is 219 and 781 out of 1000.
The time was 00:00:01.6365655 for: iteration 7
Starting timing for: iteration 8
Winrate is 254 and 746 out of 1000.
The time was 00:00:01.6362973 for: iteration 8
Starting timing for: iteration 9
Winrate is 252 and 748 out of 1000.
The time was 00:00:01.6188291 for: iteration 9
Starting timing for: iteration 10
Winrate is 219 and 781 out of 1000.
The time was 00:00:01.6260666 for: iteration 10
Starting timing for: iteration 11
Winrate is 227 and 773 out of 1000.
The time was 00:00:01.6598405 for: iteration 11
Starting timing for: iteration 12
Winrate is 231 and 769 out of 1000.
The time was 00:00:01.6268967 for: iteration 12
Starting timing for: iteration 13
Winrate is 237 and 763 out of 1000.
The time was 00:00:01.5929842 for: iteration 13
Starting timing for: iteration 14
Winrate is 231 and 769 out of 1000.
The time was 00:00:01.5624353 for: iteration 14
Starting timing for: iteration 15
Winrate is 225 and 775 out of 1000.
The time was 00:00:01.6153027 for: iteration 15
Starting timing for: iteration 16
Winrate is 231 and 769 out of 1000.
The time was 00:00:01.7203895 for: iteration 16
Starting timing for: iteration 17
Winrate is 222 and 778 out of 1000.
The time was 00:00:01.6577918 for: iteration 17
Starting timing for: iteration 18
Winrate is 209 and 791 out of 1000.
The time was 00:00:01.6415581 for: iteration 18
Starting timing for: iteration 19
Winrate is 215 and 785 out of 1000.
The time was 00:00:01.6697110 for: iteration 19
Starting timing for: iteration 20
Winrate is 242 and 758 out of 1000.
The time was 00:00:01.6217042 for: iteration 20
Starting timing for: iteration 21
Winrate is 213 and 787 out of 1000.
The time was 00:00:01.6501536 for: iteration 21
Starting timing for: iteration 22
Winrate is 238 and 762 out of 1000.
The time was 00:00:01.6388505 for: iteration 22
Starting timing for: iteration 23
Winrate is 220 and 780 out of 1000.
The time was 00:00:01.5725185 for: iteration 23
Starting timing for: iteration 24
Winrate is 226 and 774 out of 1000.
The time was 00:00:01.5928600 for: iteration 24
Starting timing for: iteration 25
Winrate is 215 and 785 out of 1000.
The time was 00:00:01.6803162 for: iteration 25
Starting timing for: iteration 26
Winrate is 224 and 776 out of 1000.
The time was 00:00:01.7173730 for: iteration 26
Starting timing for: iteration 27
Winrate is 210 and 790 out of 1000.
The time was 00:00:01.5831967 for: iteration 27
Starting timing for: iteration 28
Winrate is 227 and 773 out of 1000.
The time was 00:00:01.6209326 for: iteration 28
Starting timing for: iteration 29
Winrate is 239 and 761 out of 1000.
The time was 00:00:01.7199357 for: iteration 29
Starting timing for: iteration 30
Winrate is 209 and 791 out of 1000.
The time was 00:00:01.7053807 for: iteration 30
Starting timing for: iteration 31
Winrate is 228 and 772 out of 1000.
The time was 00:00:01.6197928 for: iteration 31
Starting timing for: iteration 32
Winrate is 217 and 783 out of 1000.
The time was 00:00:01.6192794 for: iteration 32
Starting timing for: iteration 33
Winrate is 239 and 761 out of 1000.
The time was 00:00:01.6510362 for: iteration 33
Starting timing for: iteration 34
Winrate is 241 and 759 out of 1000.
The time was 00:00:01.5823187 for: iteration 34
Starting timing for: iteration 35
Winrate is 230 and 770 out of 1000.
The time was 00:00:01.6404490 for: iteration 35
Starting timing for: iteration 36
Winrate is 234 and 766 out of 1000.
The time was 00:00:01.6085067 for: iteration 36
Starting timing for: iteration 37
Winrate is 225 and 775 out of 1000.
The time was 00:00:01.6630979 for: iteration 37
Starting timing for: iteration 38
Winrate is 251 and 749 out of 1000.
The time was 00:00:01.6128358 for: iteration 38
Starting timing for: iteration 39
Winrate is 225 and 775 out of 1000.
The time was 00:00:01.6979321 for: iteration 39
```

This is giving a whole new appreciation for policy gradients. Yeah, even though I could fiddle with the temperature hyperparameter to get better exploration, PG will move from exploration to exploitation in a smooth manner. Here the deep MC player just overfits and cannot move from there.

I am sold. PG will be my weapon of choice. Now if only I could ditch the MC updates through the state association that would be the best. I am calling it a day here. Tomorrow I will work on the state-action association.

I did really good today. I cleared up quite a lot of stuff. I am making slow, but noticeable progress even if it is just realizing that PG are actually good."