"10:20am. This was a really stupid mistake. It also means the the tests for the recurrent networks need to be repeated. The rewards were always zero so no optimization ever took place.
```
[|0.486473|]

[|0|]

[|0.9771808|]

[|6.660579|]

[|0|]

[|-0.668149|]

[|-4.460037|]

[|0|]

[|-3.229894|]

[|-21.86557|]
```

The gradients here are rather large.

No forget it. This can't be the answer. I simply do not have anything to go for here. The association idea is a bust. But now that I've fixed the hideous error in the trace direct, let me try RNNs again. My luck is good it seems."