"10:25am.
```
Starting timing for: iteration 0
Winrate is 245 and 755 out of 1000.
The time was 00:00:01.1738382 for: iteration 0
Starting timing for: iteration 1
Winrate is 298 and 702 out of 1000.
The time was 00:00:00.8745938 for: iteration 1
Starting timing for: iteration 2
Winrate is 278 and 722 out of 1000.
The time was 00:00:00.8833425 for: iteration 2
Starting timing for: iteration 3
Winrate is 308 and 692 out of 1000.
The time was 00:00:00.8105099 for: iteration 3
Starting timing for: iteration 4
Winrate is 304 and 696 out of 1000.
The time was 00:00:00.7898314 for: iteration 4
```

I am getting better win rates now. Let me try recurrence.

10:30am. I am getting nothing from recurrence. This idea is a bust as well.

10:35am. Let me go back MC updates. I want to see if I can get RNNs to work better by fiddling with the learning rates. A single hidden layer with LN works really great, but maybe I can do better than that?"

"10:40am.

```
Starting timing for: iteration 0
Winrate is 223 and 777 out of 1000.
The time was 00:00:07.4159098 for: iteration 0
Starting timing for: iteration 1
Winrate is 433 and 567 out of 1000.
The time was 00:00:08.8413571 for: iteration 1
Starting timing for: iteration 2
Winrate is 610 and 390 out of 1000.
The time was 00:00:08.6794151 for: iteration 2
Starting timing for: iteration 3
Winrate is 693 and 307 out of 1000.
The time was 00:00:08.6735623 for: iteration 3
Starting timing for: iteration 4
Winrate is 699 and 301 out of 1000.
The time was 00:00:07.9837585 for: iteration 4
Starting timing for: iteration 5
Winrate is 643 and 357 out of 1000.
The time was 00:00:04.1069369 for: iteration 5
Starting timing for: iteration 6
Winrate is 657 and 343 out of 1000.
The time was 00:00:04.0402042 for: iteration 6
Starting timing for: iteration 7
Winrate is 626 and 374 out of 1000.
The time was 00:00:03.8853135 for: iteration 7
Starting timing for: iteration 8
Winrate is 649 and 351 out of 1000.
The time was 00:00:03.9943425 for: iteration 8
Starting timing for: iteration 9
Winrate is 633 and 367 out of 1000.
The time was 00:00:03.9739528 for: iteration 9
Starting timing for: iteration 10
Winrate is 658 and 342 out of 1000.
The time was 00:00:03.9067274 for: iteration 10
Starting timing for: iteration 11
Winrate is 631 and 369 out of 1000.
The time was 00:00:03.8004685 for: iteration 11
Starting timing for: iteration 12
Winrate is 633 and 367 out of 1000.
The time was 00:00:03.8705423 for: iteration 12
Starting timing for: iteration 13
Winrate is 644 and 356 out of 1000.
The time was 00:00:03.8869292 for: iteration 13
Starting timing for: iteration 14
Winrate is 644 and 356 out of 1000.
The time was 00:00:04.0159830 for: iteration 14
```

It is indeed possible to make it work better, but then it makes a bad update and gets wrecked. Let me try the feedforward net again.

10:40am.

```
Starting timing for: iteration 0
Winrate is 266 and 734 out of 1000.
The time was 00:00:05.0852239 for: iteration 0
Starting timing for: iteration 1
Winrate is 530 and 470 out of 1000.
The time was 00:00:05.6949513 for: iteration 1
Starting timing for: iteration 2
Winrate is 665 and 335 out of 1000.
The time was 00:00:04.4912478 for: iteration 2
Starting timing for: iteration 3
Winrate is 683 and 317 out of 1000.
The time was 00:00:04.0973329 for: iteration 3
Starting timing for: iteration 4
Winrate is 730 and 270 out of 1000.
The time was 00:00:04.1801811 for: iteration 4
Starting timing for: iteration 5
Winrate is 703 and 297 out of 1000.
The time was 00:00:04.0695968 for: iteration 5
Starting timing for: iteration 6
Winrate is 736 and 264 out of 1000.
The time was 00:00:04.0338619 for: iteration 6
Starting timing for: iteration 7
Winrate is 722 and 278 out of 1000.
The time was 00:00:03.9393911 for: iteration 7
Starting timing for: iteration 8
Winrate is 722 and 278 out of 1000.
The time was 00:00:04.0880370 for: iteration 8
Starting timing for: iteration 9
Winrate is 734 and 266 out of 1000.
The time was 00:00:03.9596892 for: iteration 9
Starting timing for: iteration 10
Winrate is 742 and 258 out of 1000.
The time was 00:00:04.0931265 for: iteration 10
Starting timing for: iteration 11
Winrate is 737 and 263 out of 1000.
The time was 00:00:04.0254571 for: iteration 11
Starting timing for: iteration 12
Winrate is 708 and 292 out of 1000.
The time was 00:00:03.8206465 for: iteration 12
Starting timing for: iteration 13
Winrate is 710 and 290 out of 1000.
The time was 00:00:04.2376187 for: iteration 13
Starting timing for: iteration 14
Winrate is 742 and 258 out of 1000.
The time was 00:00:04.1107745 for: iteration 14
```

The feedforward net with another 3x decrease in learning rate (same as for the RNN) is really, really good. It is extremely stable and has good performance. I do not need to worry about exploration with it.

Seeing all my ideas turn out to be bullshit is a real eye opener for me.

Really, I should just forget about making significant algorithmic innovations on my own. I am going to change my strategy. I'd be better off looking for overlooked gems like the optimization algorithm from that dendrical microcircuits talk or putting in novel architectures like differentiable plasticity.

Given how poorly the MIRNN works, I can freely take it out of consideration. The architecture I am going to use will have to be good on everything.

I guess the next step from here would be to implement PPO perhaps? Actor critic?

Self play? Larger games like full NL Holdem? Checkpointing for NNs?

There is a lot of stuff to do. This is where everything stats. Even with standard PG, I can't say that I have nothing.

I think I am going to give RL a rest while I implement differentiable plasticity for the char RNN. Let me take out all the action-state association nonsense out and then I will go have breakfast."