I got up an hour ago and have been replying and asking question in that PPO thread. I've cleared out all of my misunderstanding and now I understand both PPO and PG completely.
I'll want to review the PG math once more though. And I want to study importance sampling today.

In addition to all of that, what was cleared up yesterday almost by accident is how the sigmoid cross entropy works. The link I posted in one of the replies was wrong, but it occurred to me what it could be used for. I once tried simplifying sigmoid CE and failed. Now I see how to do it.

I also want to rederive the softmax once more. I completely forgot the technique and want to refresh my memory of it. Also, I want to do it with the cost function as well which is something I did not do last time.

Let me start with this last one.