https://www.youtube.com/playlist?list=PL6EA0722B99332589
Right now I am taking a break since I skimmed like 200 pages of the Numerical Optimization book since the early morning. That took a bite out of me.

Damn RL. The difficulties with it are not going to be possible to solve with better models. It is going to take better optimizers.

Or achitectures that do not do backprop.

Right now, let me at least understand the basics of what second order methods are doing. After that it will be a lot easier. I'll start by watching the first lecture from the above course. Hopefully it will be easier to digest than the book.

When the time comes, I think I will implement CurveBall using Diffsharp and try it out.

But before that I'll also hit up the cortical microcircuits paper to figure out what is going on there.

I need something better. If I had something better then maybe even MC methods would do well.

It is time to resume.