The quest to master numerical optimization continues for another day.
The relentless flurry of new architectures won't be enough for RL like it was for SL. I will be necessary to go a step further for RL.

I will do what I can to make that happen. I need to increase my knowledge for the sake of that.

If I could improve the credit assignment scheme for the RL algorithms, that would open the the door to bring differentiable plasticity and the rest of those good SL architectures.