https://youtu.be/XpPvsMhxwSM?list=PL6EA0722B99332589
Let me stop here.

So I am done with the first 16 videos. I have a decent grasp of what has been covered so far and there are only 3 videos in the unconstrained optimization section left to go. I'll leave them for tomorrow.

After that I'll study information geometry for a while, though I'll also want to go through the NO book again.

When all of that is done, I'll go through the CurveBall paper again and see if anything new light up for me. I'd be surprised if it did not.

Really if there was something like a natural gradient update that is linear in complexity and space, my problems would all be solved. As it is even if I came up with something like a QN method, it would be way too expensive. Nothing except linear scaling will suffice in the regime I am working in.

This stuff in the course is definitely some brilliant math, but it is not what I need. It is just a stepping stone to better things.

Today my mind kept going back to the Baird's couterexample, and I am thinking that for bootstrapping objectives, I really should put centered layer norm in. There is not much choice. Layer norm, weight norm, batch norm those are the real novelties of modern deep learning.

I need to combine them with relative NG methods and that CurveBall thing.

That should get me somewhere. I need to improve my understanding and then take a stab at it once again. I need to understand what NNs are doing.

https://arxiv.org/abs/1610.06160
Streaming Normalization - Towards Simpler and More Biologically-plausible Normalizations for Online and Recurrent Learning

I just Google online BN and got this.

https://arxiv.org/abs/1805.10694
Towards a Theoretical Understanding of Batch Normalization

This was posted on the ML sub today and I intend to study it in depth later. Normalization schemes and optimizers are what will get me to that vaunted level where I can make a good player and move onto the next phase.

I need to keep going forward. For the past 3 years I did not really study ML in any depth. It is more like my occupation was figuring out how to make a library. Now that this is done to a degree - adding higher order AD will complicate things, I can move to really leveling up my skills.

The second part of my journey has already started.