"7:05pm. https://arxiv.org/abs/1805.11604 How Does Batch Normalization Help Optimization? (No, It Is Not About Internal Covariate Shift)
https://arxiv.org/abs/1805.11593
Observe and Look Further - Achieving Consistent Performance on Atari

7:10am. Oh, this second paper might be something.

"We argue that an algorithm needs to overcome three key challenges in order to perform well on
all Atari games. The first challenge is processing diverse reward distributions. An algorithm must
learn stably regardless of reward density and scale. Mnih et al. [13] showed that clipping rewards
to the canonical interval [−1, 1] is one way to achieve stability. However, this clipping operation
may change the set of optimal policies. For example, the agent no longer differentiates between
striking a single pin or all ten pins in BOWLING. Hence, optimizing the unaltered reward signal in a
stable manner is crucial to achieving consistent performance across games. The second challenge
is reasoning over long time horizons, which means the algorithm should be able to choose actions
in anticipation of rewards that might be far away. For example, in MONTEZUMA’S REVENGE,
individual rewards might be separated by several hundred time steps. In the standard γ-discounted
RL setting, this means the algorithm should be able to handle discount factors close to 1. The third
and final challenge is efficient exploration of the MDP."

7:15am. Damn, I am still groggy. Let me read this paper. After that come the lectures. I'll finish them today. After that will come those most IG ones. Well, unless the Deepmind paper turns out to be so particularly good that I can switch my plans again."