https://arxiv.org/abs/1412.1193 New insights and perspectives on the natural gradient method
I started this monster paper today and will move into it properly tomorrow. There is no point continuing past this point.

Yesterday I wanted to move back into programming, but during the night I realized that I want to do some research after all.

I'll guess making this choice is natural for me. I already do not think the existing methods will suffice anyway. If I implement relative natural gradient, it would obsolete PPO anyway. I really like it as its story makes sense despite it being proof of concept.

Today I realized the the KL divergence's negative Hessian is in fact also its Fisher matrix. That is not true in general. That was so confusing. I thought natural gradient was a second order method because I saw that being used.

One might imagine that this would make me favor CurveBall more as the optimization it does would be equivalent to natural gradient in the RL domain, but that is not necessarily the case.

It being natural gradient or not aside, it is true that second order methods work well on things like the Resonbrock function where first order methods struggle. When the number of parameters is small, a quadratic approximation to the target function is a lot better assumption than when the number of parameters is million.

Would approximating such huge functions with a quadratic make any more sense that standard SGD?

Furthermore, it would not necessarily be the case that in the RL domain it would be doing natural gradient learning. For it to be NG, it would have to take the expectation over all the actions rather than the one being picked.

This is not what CurveBall would be doing. With its Gauss-Newton approximation I'd get the equivalent of NG, but just on a single sample. Not only that, but it would only do a single step of optimization over its quadratic model and I might want it to be more if the gradients are getting out of control.

What I need is stability.

Relative natural gradient which operates on a local scale is the best bet for this.

It is probably the ideal use case for Spiral and I should take advantage of this opportunity to prove its worth.

I will need to educate myself to a significantly larger degree than if I was doing only first order optimization in order to get a grip on this, but that just goes with the territory. I like getting what could be permanent advantages over my competitors.

I estimate it will take me a few months to get this all under wraps. I won't put a deadline on this. I'll start work on implementing this when I feel my understanding has reached a sufficient level. I'll continue my studies until then.

One thing I will do tomorrow is email Ke Sun and ask him whether he considered replacing the matrix inversion with a few steps of SGD. I guess I'll ask him whether there is something special for recurrent nets or if I should treat it as a feedforward layer for purposes of optimization.

Well, after resolving the issue that came in, can't forget that.